model_downloader: treat SSLError as recoverable
If the TCP stream from an HTTPS server gets corrupted, this will cause
the TLS MAC check to fail, which causes an SSLError. Since such an error
is transient by nature, we should attempt to recover from it.

Also, requests.exceptions.HTTPError is a subclass of RequestException,
so remove the redundant handler for it.