model_downloader: make the specification of files to download more generic
This frees us from having to contort the meanings of "model" and "weights" to
support various frameworks and makes it possible to support more than two files
per model.

This change also makes the downloader keep the downloaded archive and all of
its contents for models downloaded as archives. This is possible now, because
the config now contains hashes for archives rather than their contents,
and therefore the archives can be retrieved from a cache (when using one).