Extract the model downloader's dependencies into a requirements file
This will make installing the dependencies easier for users and will also
help with the implementation of our CI.

The file is called requirements.in rather than requirements.txt, because
the dependencies listed within aren't pinned. One can turn it into a real
requirements.txt file using pip-compile (which is what I intend to do for
our CI).

Also, change the install command in the documentation to not require root
privileges.