downloader.py: produce less output when not printing to a terminal
With our small chunk sizes, downloading large networks produces an obscene
amount of output, which is hard to read in CI logs. Mitigate this by using
a larger chunk size when standard output is not a terminal.

Also, don't attempt to overwrite the progress line when not printing
to a terminal. It makes no sense, and it confuses programs that read the
downloader's output into thinking that it's printing one giant line.