demos: replace links to the repository with relative links
Since we've finally moved the demos to this repo, there's no longer a reason
to use absolute links to refer to other stuff in this repo. Doing so is
wrong anyway, since we can't guarantee that the master branch (or any other)
will hold any given content in the future.