tools/downloader: add support for per-model configs
Moving to per-model configs will make it easier to maintain
the model configuration and reduce the amount of information that needs
to be included in the config (since name and subdirectory are now inferred).

This change is transparent to users who don't read list_topologies.yml directly.

For compatibility the -c option is still supported and will let
users use a config in the old format. However, it is deprecated and will be
removed in a future release. I didn't provide a way to specify an alternate
tree of model configs in the new format, because such a feature doesn't seem
useful to me (and the config format is still not supposed to be stable).