tools/downloader: purge the "topology" term
So far we've been treating the words "topology" and "model" synonymously,
which is both incorrect and confusing. The user-facing documentation
(as well as the tools' names) uses "model", so update the internals to match.

Strictly speaking, "model" isn't correct either, since the things we call models
are actually families of different-precision models. However, the difference
is minor enough that I think we can ignore it.

The only occurrence of "topology" that remains is the name of the key
in the deprecated combined config file format.