tools/downloader/downloader.py: do all output through a Reporter object
There are two main reasons for this:

a) ensuring consistent formatting and output stream for all messages;
b) adding the possibility to disable all normal output when machine-readable
   output is added in the future (which I plan to do soon).

The latter is also why there is a reporter _object_ instead of a few global
functions - this object will later be initialized differently based on
command line arguments.

Errors and warnings are now printed to stderr. When machine-readable output
is added, we'll still want to display human-readable error messages, but
they'll have to go to stderr to avoid intermixing with machine-readable stdout.

Various subtypes of Requests exceptions are no longer handled, because
we now print the full traceback, which ought to contain enough details already.