tools/downloader: Document a workaround for a TLS connection issue
The official macOS Python 3.5 binaries are linked with system OpenSSL,
which doesn't support TLS 1.2. Unfortunately, some of our file sources
only support TLS 1.2 and up.

The workaround is to install 'requests[security]', which includes pyOpenSSL
(which Requests will automatically use if it's available).

Since Python 3.6, the official binaries bundle a newer version of OpenSSL,
so the issue no longer exists.