demos/tests/cases.py: avoid an out-of-bounds access in smart_classroom_demo
This demo assumes by default that the action recognition model recognizes
3 actions, but person-detection-raisinghand-recognition-0001 only recognizes 2.
This causes the demo to go out of bounds when retrieving the action
probabilities.

We could fix it by adding a -student_ac option to the test case, but a better
fix is to use the top-k mode (via the -a_top option) for this model, since:

a) in top-k mode the demo already defaults to 2 actions; and
b) this gives us better code coverage.

In top-k mode the other 3 models aren't used, so don't add cases with them.