demos/tests/cases.py: prepare for the disappearance of libcpu_extension.so
libcpu_extension.so is going to be merged into the CPU plugin in
IE 2019 R4, so don't add the -l option if it isn't present. This
will allow the test script to work with both R3 and R4. After R4
is released, we can remove the passing of the option entirely.