demos: replace vectors with arrays where possible
The impetus for this is that an internal static analysis tool gets confused
by vectors initialized with initializer lists. However, it's a good change
in its own right: by using arrays we avoid unnecessary dynamic allocation
and, in some cases, dynamic initialization.

In one case, the vector isn't necessary at all (it shadows another identical
vector), so remove it completely.