demos/mask_rcnn_demo: prevent OOB accesses to the colors array
There's no guarantee that the number of unique classes in the image is less
than the number of colors, so use a remainder operation to wrap the color
index.

Also, change the way that `class_color` is populated to prevent dependence
on evaluation order. Evaluating `class_color[class_id]` can change the size,
so the color indexes could change depending on whether the LHS or the RHS
was evaluated first. Use `emplace` to set the default value instead.