smart_classroom_demo: factor out the face recognition code into a class
This ensures that none of the face recognition components are used when
recognition is disabled, and makes it clearer which parts of the code
are responsible for face recognition.

My eventual goal is to remove CnnConfig::enabled, so that the neural
network wrappers can never exist in an uninitialized state. However,
for that to happen, the code using the action & face recognition networks
will have to be refactored as well.