Model converter: perform conversions in a deterministic order
`model_precisions` is a set, so the order of iteration over it can vary
from run to run, which can be mildly confusing. We could try to preserve
the user-specified precision order, but that's complicated and doesn't
seem all that useful. It's much simpler to just do it in alphabetical order.