smart_classroom_demo: remove the `enabled` fields from detector configs
Instead, use the null object pattern to disable a detector when
the corresponding model option is unspecified.

My original aim with this was to fix static analyzer warnings about the fields
of `ActionDetection` not being initialized (which they weren't, when `enabled`
was false). I think it's a good change in its own right, though, since it applies
the Single Responsibility Principle by separating the "detection enabled" and
"detection disabled" behavior into different classes.