Add a script that determines which CI jobs to run for a pull request
Currently, this determination is done by a similar script in an internal
repository used by the CI jobs. There are, however, a few advantages to
having this script in the main repo:

* It's easier to effect project layout changes, since we can update
  the attributes and/or the script in the same PR in which we update
  the layout, thus avoiding periods where the project is desynchronized
  from the CI logic.

* Different branches can contain different versions of the script,
  which lets us configure CI logic per-branch without special-casing
  it in the CI scripts.

* We can use Git attributes to specify most of the job triggering policy,
  which lets us easily create fine-grained rules.

* It's more transparent to external users.

Admittedly, since the actual definitions of CI jobs are still in an internal
repository, it'll still be necessary sometimes to synchronize changes to the
main and CI repos. I hope to eventually move more of the CI logic to the main
repository to rectify this.

There are five possible jobs that the script can request: "ac", "demos",
"documentation", "downloader", and "models". We don't actually have separate
"downloader" and "models" CI jobs (both types of changes are handled by the
downloader job), but I think that:

a) it's more logical to report these types of changes separately; and
b) we might eventually create a "models" job to separate the testing of models
   from the testing of the downloader itself.