downloader: add more supported precisions in preparation for the next release
In the next release, there will be quantized models based on both FP32
and FP16, so additional precision IDs will be necessary to distinguish them.

We should be able to remove the old INT1/INT8 IDs in the future,
but for now we still have models using them, so they have to stay.