mask_rcnn_demo: use the documented output of the models
The documented output is `reshape_do_2d`, not `detection_output`. The current
code has worked so far, because a layer named `detection_output` also exists
in the network; however, it's an internal layer and we shouldn't be relying
on it.

Note that currently, `detection_output` is also marked by MO as an output
(unless you use IR v7). This might, however, be a bug. An internal report
already exists for it.