Add a file with pinned versions of Accuracy Checker's test dependencies
pytest and pytest-mock are already pinned in the .in file, but they have
a boatload of dependencies, so it would be safer to pin them all.

Note that `requirements-ac-test.txt` is built from _both_ `requirements.in`
and `requirements-test.in`. This is so that if main and test dependencies
ever overlap, `requirements-ac-test.txt` will respect any restrictions
set in `requirements.in`. The test dependencies have to be installed into
the same virtual environment as the main dependencies, so we can't have any
conflicts between them.

It is annoying to have duplicated dependencies in two files, but since
they're generated, I don't think it's a big problem. One alternative would
be to just integrate the test dependencies into `requirements-ac.txt`,
but we might add some build jobs that use AC without also testing it,
and I'd rather not install the test dependencies in that case.

I also update `requirements-ac.txt` to keep the versions in both files
synchronized.