pytorch_to_onnx.py: remove --from-torchvision
This flag is redundant, as --import-module=torchvision.models has the same
effect (with the only difference being a slightly less clear error message
when the package is missing). Removing it simplifies the code.

As a side effect, this also allows using --model-param with torchvision models,
although there is currently no reason to do so.