Normalize arXiv links
Currently, some of the links to arXiv go to the abstract page and some go
to the article PDF. Normalize it so that they all go to the abstract page.
You can get to the PDF from the abstract page, but you can't do the opposite,
so the abstract page is more useful.