models/public/resnet-50-tf: convert directly from the SavedModel
No need to use an intermediate conversion step. This results in a model of
the same size, with the same accuracy metrics.

I used the non-JPEG variant of the original model, since it's slightly
smaller and easier to convert (you can use the original placeholder instead
of sticking a new one in the middle of the graph).