run_tests.py: download/convert all models before starting testing
The current approach with downloading/converting from each demo's `models.lst`
doesn't scale well for several reasons:

* If multiple demos use the same model, we end up downloading/converting it
  multiple times. Even though the downloading is cached, the cache might be
  in a remote shared folder. And, of course, the conversion can't be cached at all.

* There might not be cases for all models in models.lst, which leads to wasted
  downloading time.

* Models used in different demos can't be converted in parallel.

So now we gather the list of all models used in all test cases and get them
all in advance. This does have the disadvantage of no longer ensuring that each demo's
models.lst contains all models we test it with; I plan to fix that shortly.

WRT precisions, I just get the list of all precisions used in the tests and fetch those.
It would potentially be more efficient to run the downloader individually on every model
specifying the needed precisions for each; however, right now that wouldn't improve anything
and it would be a lot more complicated.