ci/update-requirements.py: enable usage with pip-tools >=4.5.0
pip-tools has recently changed the output format by adding the source file
paths to the comments. This causes problems for us, because some of our
source files are from the OpenVINO toolkit directory, whose path changes
depending on the version and the install directory. Using that path as-is
would mean leaking local file system paths, as well as meaningless diffs
when upgrading to newer OpenVINO versions.

Fortunately, we already use a script, which means we can just postprocess
the unwanted parts away.

I also used this postprocessing to add a comment directing developers
to the update script.