Model downloader: print progress even if zero time passed
It is possible for `duration` to be 0 if a chunk has already been received
completely by the time we got to the chunk loop. In this case, we can't
calculate the speed, but we can still output all the other progress
information, so do so.

This is particularly important for very short files which entirely fit in
one chunk, since previously there wouldn't be any progress information for
them at all.

I also removed a couple of unnecessary casts in the affected code.