tools/downloader: make the downloader interruptible in parallel mode
KeyboardInterrupt is only raised in the main thread of the application,
so if you press Ctrl+C while the downloader is working, any jobs that
have already started will run to completion before the program is terminated,
which is a bad user experience. To fix this, make the jobs manually check
if they should stop before doing anything that might take some time
(such as downloading the next chunk), and tell every job to stop once
an exception is raised.

Note that Model Converter doesn't need this kind of help, because its
jobs consist solely of launching other programs. Ctrl+C is delivered to
all running programs, so any subprocesses running when it's pressed will be
terminated, allowing the conversion jobs to finish quickly.