security_barrier_camera_demo: fix stack-use-after-scope
There was a situation when the context was destroyed and the worker still lived with its ReborningVideoFrames referring to the destroyed context and trying to reach the context in ~ReborningVideoFrames().
I could clear() a worker’s set of tasks on stop() but that would result in (almost) infinite loop because ReborningVideoFrames would create themselves again. That would look like the demo hangs for a few seconds on Esc before exit.
Another solution was to store a smart ptr in ReborningVideoFrame to the context instead of reference. But it would make things more interlinked or I would need to mess with casting weak_ptr<Context> to shared_ptr.