downloader: disable compression when resuming a download
Requests enables compression by default. However, it's not a good idea to do
that when making a range request, because if the server decides to use
compression, then it'll return a range from the _compressed_ file, which is
useless (you can't decompress a chunk from the middle of a compressed file).

Because of that, I suspect that most servers already either disable compression
or don't honor range requests, but just in case, disable it on our side too.