Enable input channel reversal for deeplabv3
The Model Optimizer issue that prevented us from doing this has been
fixed at some point, so there's nothing preventing us from doing it.

This makes this model's input consistent with most others, and makes it
so that people can use it with our segmentation demo without modifying
the demo (or losing accuracy).