demos: turn common into a static library
This has several benefits:

* Its files can be seen in IDEs;
* It allows us to move code from header files to implementation files,
  reducing build times and header dependencies.
* We can use CMake's target mechanism to avoid duplicating the path
  to the headers.

In order to keep the diff minimal, I only made one implementation file so far.
I chose to split args_helper.hpp, because it indirectly included windows.h,
which is huge, so it's always nice to limit where it's included.
Later on, implementation files for other headers ought to be created as well.