downloader: don't delete the whole output directory if the download fails
The purpose of deleting it is to make sure the user doesn't end up with an
incomplete or unverified file. However, deleting the whole directory is
overkill, and can lead to nasty surprises if the user sets an existing
directory as the output directory.

Instead, just delete the file we were trying to download. This means we can
leave the user with an incomplete _set_ of files, but that doesn't seem like
a significant issue.