Add a script to prepare documentation for inclusion into OpenVINO docs
The layout file and page IDs generated by the script were previously
maintained by the OpenVINO documentation team. Having them generated
automatically lifts that burden off their shoulders and also makes it
possible for us to autogenerate other aspects of the documentation
in the future.

In the generated layout file, the titles of some pages are overridden.
This is so that they match the titles in the table of contents from
the previous releases of OpenVINO. Ideally, we should align the titles
in the layout with the titles in the pages themselves, but I'm not yet
sure which ones are better, so I'm deferring that for later.

About the requirements file: I used a pretty generic name for it
(requirements-documentation), because I expect that other
documentation-related scripts (e.g. a model index checker) might be added
later. I'll add a version with pinned dependencies
(requirements-documentation.txt) when I integrate this script into CI.