get-jobs-for-changes.py: request a models.lst check when a model is deleted
Occasionally we merge a PR that deletes a model and then demo tests start
failing, because one of the demos depended on it. This change is a step
towards preventing that from happening.

Right now, there isn't anything actually connected to this "models_lst"
identifier. I plan to add a check to the demo job that just runs the info
dumper on every models.lst file in the repository, and to trigger that
check when the "models_lst" job is requested.

I could just set it to run the full demo job when a model is deleted, but that
seems like an overkill; the demo tests take a while, and a model can only be
used in tests if it's listed in models.lst, thus checking models.lst is
sufficient.