Add a script to do basic checks
The idea is that these should be reasonably quick checks that we can do
for every pull request.

I disable the yamllint newline check, because I want the basic checks to be
runnable on Windows, and this check fails on Windows (because Git uses native
line endings when checking files out).