Bump minimal Python version to 3.6
Python 3.5 is EOL and OpenVINO doesn't support it anymore, so I don't think
we need to support it either.

This removes the need to cap the PyTorch version, so don't do that
anymore.

We also technically don't need the workaround with the forced atomicwrites
requirement in requirements-test.in anymore, since now it should be possible
to regenerate the requirements-*.txt files on Windows. I don't want to
only be able to do it to Windows, though, so I'd rather keep the workaround.