Fix issues found by yamllint
In case of the deleted comments, yamllint was complaining about indentation,
and while I could've fixed it, we already have reference metric values in
the documentation, so I don't think we need to duplicate them in the config.

In the case of googlenet-v3-pytorch, the values in the config were different
from the ones in the documentation. I chose to replace the values in the doc
with the ones from the config, because:

1) The torchvision documentation lists the same values;
2) I ran AC on this model, and the results are closer to these values than
   the ones that are currently listed in the doc.