faceboxes-pytorch: use human-friendly names for outputs (#1629)
I don't see any advantage in using numeric names, and it has a disadvantage:
PyTorch fails to export the model to ONNX if opset is set to version 11
(presumably numeric names are disallowed). So replace them with descriptive
names.

The demo doesn't seem to care what the names are, so I'm not modifying it.