yolo-v2-tiny-tf: use a pre-convert script to convert to frozen graph
One thing to note is that the convert.py script from the keras-YOLOv3-model-set
repository can, with minimal persuasion, save the model directly in
the SavedModel format, which would make the conversion procedure much simpler
and require fewer patches.

The only drawback is that it requires TensorFlow 2, and since MO still supports
TensorFlow 1, I'm not going to do it just yet. But it is something to keep
in mind for later.