Model Converter: add support for custom pre-convert scripts
This allows adding models which can't be downloaded directly in a form
usable by Model Optimizer. Such models can be downloaded in whatever
form they're available in and then pre-processed with a custom script.

Previously, we have dealt with this by doing the pre-conversion step
offline, uploading the results to download.01.org and publishing a recipe
on how to recreate it.

* The recipe is extra documentation that must be written by model contributors.
* The recipes cannot be validated by CI. There's no assurance that the recipe
  is correct or that the original files it references are still available.
* The users (or even other developers) have no assurance that the file on 01.org
  was actually produced by the recipe.
* The need to upload files to 01.org creates a bottleneck in the contribution
  process, especially for external contributors.

Pre-convert scripts eliminate these drawbacks. However, to be fair, they have
drawbacks of their own:

* Generally, the more steps in the conversion process, the more fragile it is.
  The pre-convert script could fail due to a regression in a dependency, or some
  unforeseen portability issue.

* The pre-conversion process may require extra dependencies, which would
  otherwise not be required.

Because of these issues, I expect pre-convert scripts to only be used in cases
where pre-conversion doesn't require any dependencies beyond what is already
required to convert the model to IR. This maximizes convenience to the users
and minimizes risk.