Add missing override qualifiers
This is a) a good thing to do, and b) fixes a failure to build on macOS,
because Clang has -Winconsistent-missing-override as part of -Wall, and so,
thanks to -Werror, fails to compile demos/multi_channel/common/input.cpp.

I'd like to add -Wsuggest-override to compiler options, but we still have
to support CentOS 7, and the version of GCC in there doesn't have it.