downloader: update the list of representative models
mobilenet-v2-pytorch can no longer be used to test external module support,
because we've recently replaced it with the equivalent torchvision model.
However, it _can_ be used to test torchvision support, and it's a better
representative for that than resnet-50-tf, due to being much smaller.

Its replacement will be faceboxes-pytorch, because it's very small and yet
covers external modules, Google Drive downloads and even custom model
parameters.