ci/get-jobs-for-changes.py: be more careful with per-model Python files
A Python file will not necessarily be in the same directory as a model.yml
file. It could be in a subdirectory or in a parent directory (e.g. in
a composite model's directory). Therefore, we can't just take the parent
directory's name as the name of the model; we need to search for
the corresponding model.yml file.

I made it so that we search for the model.yml file in all parent directories;
this will work correctly if a Python file in a subdirectory of a model
directory is changed. However, if a Python file in a composite model's
directory is changed, that change will be ignored. It's probably a good idea
to make it so that such a change triggers the build for all subcomponents
of the composite model; however, there's currently no foolproof way to identify
a composite model, so I'll put that off for later.

In addition, don't trigger a models.lst check if a Python file is removed;
that only makes sense if a model.yml file is removed.