demos: Remove mandatory -Werror
Having -Werror on by default means that the build can get broken if the code
is built with a version of the compiler that we didn't test with. This isn't
just theoretical - it has already happened (see 66aa1d1a and 948b4d8f).

-Werror is useful, but for developers rather than users. I plan to add it
as an extra option in our CI, so that we can benefit from it ourselves without
risking breaking the build for our users.

Also, remove the Windows equivalent, which is not enabled by default, but
hidden behind an undocumented and unnecessary option. -DCMAKE_CXX_ARGS=/WX
accomplishes the same effect as TREAT_WARNING_AS_ERROR, and requires no
additional code.