mask_rcnn_demo: skip zero-size boxes
These aren't normally supposed to occur, but there's a currently a bug
in OpenVINO master where the GPU plugin produces invalid boxes when inferring
with mask_rcnn_resnet101_atrous_coco. This change works around that (and,
in general, makes the demo more resilient to misbehaving networks).

Remove the clipping of box_width and box_height, because clipping them from
below is made redundant by this change, while clipping them from above isn't
necessary in the first place - the corners are clipped, so the box can't be
bigger than the original image.