Convert ctdet_coco_dlav0_384 from the original files
Turns out we don't actually need any of the CenterNet requirements
to convert this model, and in fact, only need 3 files from the source
repository.

I have not retained the change to dlav0.py that was included in
pytorch-onnx.patch, because I believe we no longer need it. The only
effect of that change was to ensure deterministic order of model outputs,
but we now depend on Python 3.6+, in which dictionaries are deterministic
already.

Since I was modifying this model anyway, I took the opportunity to give
its outputs human-friendly names.

Other than the change in names, the IRs generated by the new config are
identical.

One big downside of this change is that the PyTorch weights file is strangely
three times as big as the ONNX file. I guess you win some, you lose some.