googlenet-v1-tf: convert from original files by using a pre-convert script
Pretty much the same thing as ba2ddb13, but for googlenet-v1-tf.

This doesn't change the XML file in the resulting IR (apart from layer names),
but the BIN file does change, and I don't know why. However, I retested
the accuracy of the IR (on CPU with FP32 precision), and it exactly matches
what's listed in the documentation, so I assume the differences aren't
significant.