Switch object_detection_demo and common/models to pathlib
Using pathlib results in more compact and correct code compared to os.path
(or just plain string manipulation), so I think it's about time we migrated
the demos to it.

object_detection_demo is as good a place to start as any, and by updating
common/models I create an incentive for future demos to use pathlib as well. :)