docs: replace HTML links with Markdown links
Markdown links are shorter and more readable than HTML links, so they should
be used whenever possible.

The HTML links were originally used because Doxygen didn't process external
links to Markdown files correctly. This was fixed in Doxygen 1.8.16, and
the OpenVINO documentation is now built using a newer version, so that bug
is no longer relevant.