ci/prepare-documentation.py: remove the hardcoded task type list
I originally added the list so that if a model with a new task type
is added, the prepare-documentation script crashes and so the contributor
has to consider what the human-readable description of the task type should
be.

However, given that:

1. for almost all task types, the default string transformation yields a good
   description, and

2. the cost of getting the description wrong is just an awkwardly-named section
   in the documentation TOC,

I don't think that the cost of keeping an extra copy of the task type list
is justified. Instead we should just instruct contributors to name task types
so that the default transformation gives a good result, or, if that's not
possible, to add an override.