multi_camera_multi_target_tracking_demo: use SimpleNamespace instead of dict in the configs
This makes the access syntax for configuration elements more natural (albeit
it forces the use of vars() to transform attributes into function arguments).

Change the config loading code to produce SimpleNamespaces too, so that
the main config object can be accessed the same way as its children.

To simplify the code, make analyzer.enable a mandatory attribute. There
seems to be no real reason to have it be optional, since it's already
present in both example configs, and a user who doesn't want to specify
analyzer settings already has the option of setting analyzer to None.