Enable flake8-comprehensions warning C408
Fix the last remaining violations.

I would've preferred to turn the profiles in speech_recognition_demo into
instances of types.SimpleNamespace (instead of changing the initialization
syntax), but they have to be compatible with what yaml.safe_load returns,
which is a dict.

Granted, it would be possible to convert the result of yaml.safe_load into
a namespace, but then there would need to be error handling for the case
that the keys are not strings; plus if the profile ever needs to contain
nested dicts, those would need to be converted too. All in all, it's
possible, but it looks like enough of a mess that I don't want to get into
it. :-)